% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{as.data.frame.ICCier_datagen}
\alias{as.data.frame.ICCier_datagen}
\title{Convert datagen to data.frame}
\usage{
\method{as.data.frame}{ICCier_datagen}(d)
}
\arguments{
\item{d}{datagen output.}
}
\value{
data.frame.
}
\description{
Convert datagen to data.frame
}
\keyword{internal}
