% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{convert_datagen}
\alias{convert_datagen}
\title{Convert datagen to data frame. (Deprecated)}
\usage{
convert_datagen(d)
}
\arguments{
\item{d}{datagen output.}
}
\value{
data.frame.
}
\description{
Convert datagen to data frame. (Deprecated)
}
\keyword{internal}
