% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{datagen.formula}
\alias{datagen.formula}
\title{Generate data for testing.}
\usage{
datagen.formula(formula, n, K, beta, gamma, eta, cor_structure, ...)
}
\arguments{
\item{formula}{Formula. See \code{\link{ICCier}}.}

\item{n}{Number of observations per group.}

\item{K}{Number of groups.}

\item{beta}{Location model fixed coefficients. Q_L2xQ_L1.}

\item{gamma}{Scale model fixed coefficients. P_L2xP_l1.}

\item{eta}{Between-group coefficients. R_L2x(Q_l1 + P_l1).}

\item{cor_structure}{3-length vector describing correlation matrix of REs.}
}
\value{
ICCier_dataget object. List of params, meta, and data.
}
\description{
Generates data for use in testing.
Takes a formula for the model spec (See \code{\link{ICCier}}), returns list of params, metadata, and stan data.
}
\keyword{internal}
