% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{datagen}
\alias{datagen}
\title{Generate data for testing.}
\usage{
datagen(n, K, beta, gamma, eta, cor_structure)
}
\arguments{
\item{n}{Number of observations per person.}

\item{K}{Number of persons.}

\item{beta}{1x1 matrix representing the global mean of the outcome.}

\item{gamma}{QxQ_random matrix representing the effects of Q level 2 variables on the Q_random level 1 scale parameters.}

\item{eta}{Qx(Q_random+1) matrix representing the effects of Q level 2 variables on the mean variance and Q_random level 1 scale variances.}

\item{cor_structure}{3-length vector describing the correlation matrix of random effects.}
}
\value{

}
\description{
Generates data for use in testing.
}
