% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{.datagen}
\alias{.datagen}
\title{Datagen helper.}
\usage{
.datagen(spec)
}
\arguments{
\item{spec}{List containing params, meta, and data. See details.}
}
\value{
Completed datagen list, containing params, meta, and data.
}
\description{
The spec consists of everything but the random effects and outcome.
Specifically, the spec should be a list (params, meta, data).
Params should contain the beta, gamma, eta, Omega, betaGamma_random_cor_L matrices.
Meta should contain all dimensions (e.g., n, K, N, P_l1, etc), the group_map, outcome name, and formula.
This is used for both this function, and dataframe converters, to fill in variable names.
Data should contain fixed design matrices and other stan data. \code{.datagen} is responsible for generating RE_z,RE,group-specific, and y.
}
\keyword{internal}
