% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{.formulate}
\alias{.formulate}
\title{Parses formula list into standard Formula}
\usage{
.formulate(formList)
}
\arguments{
\item{formList}{List containing between, within, location, and group formulas.}
}
\value{
Formula.
}
\description{
Takes list of formulas, converts into the standard formula.
E.g., \code{list(outcome = y, group = subjectID, between ~ age, within ~ day|1, mean ~ 1|1)}
}
\keyword{internal}
