% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.get_random_effect_samples}
\alias{.get_random_effect_samples}
\title{Extracts random effect samples.}
\usage{
.get_random_effect_samples(object, draws = NULL)
}
\arguments{
\item{object}{ICCier object.}
}
\value{
Samples of RE matrices: Arrays with dimensions KxP_l1xdraws.
}
\description{
gamma_group and beta_group are returned.
However, for prediction, we need the group-specific random effects.
This function takes the object and returns the computed REs.
}
\details{
group_gamma = x_sca_l2%*%gamma + group_gamma_random
}
\keyword{internal}
