% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.get_random_effect_z_samples}
\alias{.get_random_effect_z_samples}
\title{Extract standardized RE}
\usage{
.get_random_effect_z_samples(object, draws = NULL)
}
\arguments{
\item{object}{ICCier object}

\item{draws}{Number of draws to use from posterior samples for prediction. Default: All of them.}
}
\value{
Kx(P_l1 + 1)xS array of standardized REs.
}
\description{
log(sds) = x_sca_l2%*%eta
RE = RE_z%\*%t(diag(sds)%\*%L_cor)
RE_z = RE %*% solve(t(diag(sds)%*%L_cor))
}
\keyword{internal}
