% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{.omegagen}
\alias{.omegagen}
\title{Generate Omega matrix}
\usage{
.omegagen(cor_structure, P_l1, Q_l1)
}
\arguments{
\item{cor_structure}{3-length Vector of (loc-loc, loc-sca, sca-sca) correlations.}

\item{P_l1}{Number of random scale effects}

\item{Q_l1}{Number of random loc effects}
}
\value{
Correlation matrix
}
\description{
Generates Omega matrix from a cor_structure specification
}
\keyword{internal}
