% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parses formula using Formula.}
\usage{
.parse_formula(formula, data, predict = FALSE)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{predict}{Logical. Default: FALSE. If True, y is ignored, and the data requires grouping variable and predictors.}
}
\value{
List containing stan data, group mapping, and the model frame.
}
\description{
Parses formula using Formula.
}
\keyword{internal}
