% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_matrix}
\alias{.print_matrix}
\title{Generate character matrix for summary}
\usage{
.print_matrix(object, digits, param, ...)
}
\arguments{
\item{object}{summary.ICCier object}

\item{digits}{digits}

\item{param}{param (beta,gamma,eta,icc)}

\item{...}{Arguments passed to format.}
}
\description{
Generate character matrix for summary
}
\keyword{internal}
