% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_table}
\alias{.print_table}
\title{Generate table for summary}
\usage{
.print_table(object, digits, param, ...)
}
\arguments{
\item{object}{summary.ICCier object}

\item{digits}{digits}

\item{param}{param (beta,gamma,eta,icc)}

\item{...}{Arguments passed to format.}
}
\description{
Creates coefficient matrix ala lm and others.
}
\keyword{internal}
