% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.ICCier}
\alias{fitted.ICCier}
\title{ICCier method to extract ICC values.}
\usage{
\method{fitted}{ICCier}(object)
}
\arguments{
\item{object}{ICCier object}

\item{...}{Not currently used.}
}
\value{
TBD
}
\description{
ICCier method to extract ICC values.
}
