% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_df}
\alias{generate_df}
\title{Generate data frame for testing.}
\usage{
generate_df(n, K, beta, gamma, eta, cor_structure)
}
\arguments{
\item{n}{n}

\item{K}{K}

\item{beta}{beta}

\item{gamma}{gamma}

\item{eta}{eta}

\item{cor_structure}{cor_structure}
}
\value{

}
\description{
Uses same as \code{\link{datagen}}, but outputs a data frame.
}
