% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{is.formula}
\alias{is.formula}
\title{Tests if object is formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical. TRUE if formula
}
\description{
Tests if object is formula
}
\keyword{internal}
