% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstantools_defs.R
\name{log_lik.ICCier}
\alias{log_lik.ICCier}
\title{ICCier method for pointwise log-likelihood}
\usage{
\method{log_lik}{ICCier}(object, ...)
}
\arguments{
\item{object}{ICCier object}

\item{...}{Not used.}
}
\value{
S by N matrix of log likelihoods.
}
\description{
ICCier method for pointwise log-likelihood
}
\keyword{internal}
