% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{nsamples.ICCier}
\alias{nsamples.ICCier}
\title{ICCier method for extracting the number of posterior samples.}
\usage{
\method{nsamples}{ICCier}(object, ...)
}
\arguments{
\item{object}{ICCier object.}

\item{...}{Not used.}
}
\value{
Real value. Number of posterior samples stored (post-warmup).
}
\description{
ICCier method for extracting the number of posterior samples.
}
\keyword{internal}
