% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{posterior_interval.ICCier}
\alias{posterior_interval.ICCier}
\title{ICCier Method for posterior intervals.}
\usage{
\method{posterior_interval}{ICCier}(object, prob = 0.95, pars, ...)
}
\arguments{
\item{object}{ICCier object.}

\item{prob}{A number p (0 < p < 1) indicating the desired probability mass to include in the intervals.}

\item{pars}{Parameter type in \code{c('')}}

\item{...}{Not used.}
}
\value{
Matrix with two columns (lower and upper) and one parameter per row.
}
\description{
ICCier Method for posterior intervals.
}
\keyword{internal}
