% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{posterior_predict.ICCier}
\alias{posterior_predict.ICCier}
\title{ICCier method to predict new ICC values.}
\usage{
\method{posterior_predict}{ICCier}(object, data, ...)
}
\arguments{
\item{object}{ICCier object}

\item{data}{Data to predict from.}

\item{...}{}
}
\value{
TBD
}
\description{
ICCier method to predict new ICC values.
}
\keyword{internal}
