% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ICCier}
\alias{predict.ICCier}
\title{ICCier method to predict new ICC values.}
\usage{
\method{predict}{ICCier}(object, data, ...)
}
\arguments{
\item{object}{ICCier object}

\item{data}{Data to predict from. If NULL, calls \code{\link{fitted.ICCier}} on fit data.}

\item{...}{Not currently used.}
}
\value{
TBD
}
\description{
ICCier method to predict new ICC values.
}
