% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ICCier}
\alias{print.ICCier}
\title{Print method for ICCIer object.}
\usage{
\method{print}{ICCier}(x, ...)
}
\arguments{
\item{x}{ICCier object.}

\item{...}{Further arguments to \code{print}.}
}
\description{
Print method for ICCIer object.
}
\keyword{internal}
