% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{ranef}
\alias{ranef}
\title{Extract Random Effects}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{any fitted object from which random effects estimates can be extracted.}

\item{...}{additional arguments}
}
\value{
Depends on method function.
}
\description{
Extract Random Effects
}
\keyword{internal}
