% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.ICCier}
\alias{summary.ICCier}
\title{Summary method for ICCier objects.}
\usage{
\method{summary}{ICCier}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An ICCier object.}

\item{prob}{Probability interval (default: .95)}

\item{...}{Currently not used.}
}
\value{
TBD
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics.
}
