% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsm.R
\name{ICCier}
\alias{ICCier}
\title{Run MELSM model.}
\usage{
ICCier(formula, data, ...)
}
\arguments{
\item{formula}{Formula representing the model. See details.}

\item{data}{Data frame containing all variables.}

\item{...}{Arguments passed to \code{\link[rstan]{sampling}}.}
}
\value{
ICCier object. List containing the model formula, data, stan_data, model fit, and mapping between original ID and numeric ID.
}
\description{
Runs the MELSM model and returns an ICCier object.
}
\details{
\code{ICCier} uses the mixed effects location scale model to estimate an intercept-only
location model with random effect of person.
The within-person variances are log-linearly modelled from a set of observation-level and
person-level predictors, with coefficients \eqn{\gamma}.
The between-person variances are also log-linearly modelled from a set of person-level predictors,
with coefficients \eqn{\eta}.

The formula syntax is as follows:

\code{outcome | personID ~ Level_1_formula | Level_2_formula}

The model is implemented in a 'maximal' manner, meaning \emph{all} level 1 effects are assumed to randomly vary and correlate.
Moreover, \emph{all} level 2 variables predict each level 1 parameter.
This means you \emph{should not include cross-level interaction terms}, because they are implicit in the model formulation.
Level 1 or Level 2 interaction terms may be included.

For now, the Level 2 formula predicts both the level 1 scale parameters, as well as the level 2 random effect variances.
If you wish to only have person-specific IICs, use \code{1} as the level 1 formula (intercept-only).
}
