% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{loo.ICCier}
\alias{loo.ICCier}
\title{LOO method for computing leave-one-out (LOO).}
\usage{
\method{loo}{ICCier}(object, ...)
}
\arguments{
\item{object}{ICCier object.}
}
\value{
Loo object. See \code{\link[loo]{loo}}.
}
\description{
LOO method for computing leave-one-out (LOO).
}
\keyword{internal}
