% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ICCier}
\alias{predict.ICCier}
\title{ICCier method to predict new ICC values.}
\usage{
\method{predict}{ICCier}(object, newdata = NULL, draws = NULL,
  summary = TRUE, prob = 0.95, inc_group = TRUE, occasion = NULL,
  ...)
}
\arguments{
\item{object}{ICCier object.}

\item{newdata}{Data to predict from. If NULL, calls \code{\link{fitted.ICCier}} on fit data.}

\item{draws}{Number of draws to use from posterior samples for prediction. Default: All of them.}

\item{summary}{Logical. Whether to return summary (mean, intervals) of ICCs (TRUE), or posterior samples (FALSE)}

\item{prob}{A number p (0 < p < 1) indicating the desired probability mass to include in the intervals.}

\item{inc_group}{Logical. Whether to include the grouping variable with the estimates.}

\item{occasion}{Default: NULL. Vector representing the occasion number. One value per row in newdata. For estimating composite score reliability. If unspecified, set to NULL (default), and raw score ICCs are estimated.}

\item{...}{Not currently used.}
}
\value{
If \code{summary=TRUE}, then the mean and \code{prob}\% intervals are returned for each observation in the model frame.
If \code{summary=FALSE}, an S by N matrix containing the S posterior samples for N observations.
}
\description{
Predicts new ICC values.
}
\details{
If the grouping variable column is not included in \code{newdata}, then \code{predict} only predicts
from the fixed values.
This is \emph{not recommended}.
If the grouping variable column is present in \code{newdata}, but a value is either \code{NA} or not previously fit,
then \code{predict} marginalizes over the entire random effects distribution.
Finally, if the grouping variable is present and known, then \code{predict} marginalizes over the
uncertainty in that person's (standardized, orthogonalized) random effects.

We recommend including the grouping column, even if the grouping identifier for the predicted case
is unknown. This ensures that uncertainty in the predicted ICC includes the uncertainty of that
case's random effects.
}
