% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.ICCier}
\alias{print.summary.ICCier}
\title{Print method for ICCier summaries.}
\usage{
\method{print}{summary.ICCier}(object, ...)
}
\arguments{
\item{object}{Output of \code{summary(ICCierObject)}.}

\item{...}{Further arguments to \code{print}.}
}
\value{
Invisibly returns summary object.
}
\description{
Print method for ICCier summaries.
}
