% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.ICCier}
\alias{summary.ICCier}
\title{Summary method for ICCier objects.}
\usage{
\method{summary}{ICCier}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An ICCier object.}

\item{...}{Currently not used.}
}
\value{
ICCier summary object. List containing:
\describe{
\item{formula}{Model formula}
\item{prob}{The specified prob argument}
\item{estimate}{List of matrices of mean posterior point estimates}
\item{ci}{List containing the lower (L) and upper (U) interval estimates for each parameter matrix.}
\item{meta}{Meta-data including diagnostics, number of iterations, number of chains, total number of observations, and number of groups (persons)}
}
}
\description{
Summaries of parameter estimates and MCMC convergence diagnostics.
}
